chore: clean up project
