+++ src/app.module.ts
@@ -1,5 +1,3 @@
-// app.module.ts
-
 import { Module } from '@nestjs/common';
 import { AppService } from './app.service';
 import { AppController } from './app.controller';
