+++ src/modules/cat/cat.service.ts
@@ -1,5 +1,5 @@
-import { AutoInjectable } from 'src/auto/decorator/auto.injectable.decorator';
 import { Cat } from './cat.interface';
+import { AutoInjectable } from '../../auto/decorator/auto.injectable.decorator';
 
 @AutoInjectable()
 export class CatService {
